backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
requests==2.31.0
Pillow==10.2.0
django-ratelimit==4.1.0
